+++ CHANGELOG.md
@@ -138,6 +138,10 @@
     FIXES [#6972](https://github.com/microsoft/Microsoft365DSC/issues/6972)
 * TeamsOnlineVoiceMailPolicy
   * [BREAKING CHANGE] Changed the type of `MaximumRecordingLength` from String to Int.
+* M365DSCModuleMgmt
+  * Added the parameter `UsePowerShellGet` to `Update-M365DSCDependencies` as an override
+    fallback if `Install-PSResource` is not available or not working.
+    FIXES [#7003](https://github.com/microsoft/Microsoft365DSC/issues/7003)
 * M365DSCPermissions
   * Removed internal function `Update-M365DSCResourcesSettingsJSON`.
 * M365DSCReport